TaijaPlotting

Documentation for TaijaPlotting.

TaijaPlotting.PlotIngredients

Plots.areaplot

Plots.bar

Plots.contourf

RecipesBase.plot

RecipesBase.plot

TaijaPlotting.animate_path

TaijaPlotting.embed_path

TaijaPlotting.generate_lims

TaijaPlotting.get_names

TaijaPlotting.plot_state

TaijaPlotting.set_up_plots

TaijaPlotting.PlotIngredients — Type

A container used for plotting.

Plots.areaplot — Method

Plots.areaplot(
    conf_model::ConformalProbabilisticSet, fitresult, X, y;
    input_var::Union{Nothing,Int,Symbol}=nothing,
    kwargs...
)

A Plots.jl recipe/method extension that can be used to visualize the conformal predictions of any fitted conformal classifier. Using a stacked area chart, this function plots the softmax output(s) contained the the conformal predictions set on the vertical axis against an input variable X on the horizontal axis. In the case of multiple input variables, the input_var argument can be used to specify the desired input variable.

Plots.bar — Method

Plots.bar(conf_model::ConformalModel, fitresult, X; label="", xtickfontsize=6, kwrgs...)

A Plots.jl recipe/method extension that can be used to visualize the set size distribution of a conformal predictor. In the regression case, prediction interval widths are stratified into discrete bins. It can be useful to plot the distribution of set sizes in order to visually asses how adaptive a conformal predictor is. For more adaptive predictors the distribution of set sizes is typically spread out more widely, which reflects that “the procedure is effectively distinguishing between easy and hard inputs”. This is desirable: when for a given sample it is difficult to make predictions, this should be reflected in the set size (or interval width in the regression case). Since ‘difficult’ lies on some spectrum that ranges from ‘very easy’ to ‘very difficult’ the set size should vary across the spectrum of ‘empty set’ to ‘all labels included’.

Plots.contourf — Method

Plots.contourf(conf_model::ConformalModel,fitresult,X,y;kwargs...)

A Plots.jl recipe/method extension that can be used to visualize the conformal predictions of a fitted conformal classifier with exactly two input variable. Data (X,y) are plotted as dots and overlaid with predictions sets. y is used to indicate the ground-truth labels of samples by colour. Samples are visualized in a two-dimensional feature space, so it is expected that X $\in \mathcal{R}^2$. By default, a contour is used to visualize the softmax output of the conformal classifier for the target label, where target indicates can be used to define the index of the target label. Transparent regions indicate that the prediction set does not include the target label.

Target

In the binary case, target defaults to 2, indexing the second label: assuming the labels are [0,1] then the softmax output for 1 is shown. In the multi-class cases, target defaults to the first class: for example, if the labels are ["🐶", "🐱", "🐭"] (in that order) then the contour indicates the softmax output for "🐶".

Set Size

If plot_set_size is set to true, then the contour instead visualises the the set size.

Univariate and Higher Dimensional Inputs

For univariate of multiple inputs (>2), this function is not applicable. See Plots.areaplot(conf_model::ConformalProbabilisticSet, fitresult, X, y; kwargs...) for an alternative way to visualize prediction for any conformal classifier.

RecipesBase.plot — Method

Plots.plot(
    conf_model::ConformalInterval, fitresult, X, y;
    kwrgs...
)

A Plots.jl recipe/method extension that can be used to visualize the conformal predictions of a fitted conformal regressor. Data (X,y) are plotted as dots and overlaid with predictions intervals. y is plotted on the vertical axis against a single variable X on the horizontal axis. A shaded area indicates the prediction interval. The line in the center of the interval is the midpoint of the interval and can be interpreted as the point estimate of the conformal regressor. In case X is multi-dimensional, input_var can be used to specify the input variable of interest that will be used for the horizontal axis. If unspecified, the first variable will be plotting by default.

RecipesBase.plot — Method

Plots.plot(
    ce::CounterfactualExplanation;
    alpha_ = 0.5,
    plot_up_to::Union{Nothing,Int} = nothing,
    plot_proba::Bool = false,
    kwargs...,
)

Calling plot on an instance of type CounterfactualExplanation returns a plot that visualises the entire counterfactual path. For multi-dimensional input data, the data is first compressed into two dimensions. The decision boundary is then approximated using using a Nearest Neighbour classifier. This is still somewhat experimental at the moment.

Examples

# Search:
generator = GenericGenerator()
ce = generate_counterfactual(x, target, counterfactual_data, M, generator)

plot(ce)

TaijaPlotting.animate_path — Function

animate_path(ce::CounterfactualExplanation, path=tempdir(); plot_proba::Bool=false, kwargs...)

Returns and animation of the counterfactual path.

Examples

# Search:
generator = GenericGenerator()
ce = generate_counterfactual(x, target, counterfactual_data, M, generator)

animate_path(ce)

TaijaPlotting.embed_path — Method

embed_path(ce::CounterfactualExplanation)

Helper function that embeds path into two dimensions for plotting.

TaijaPlotting.generate_lims — Method

generate_lims(x1, x2, xlims, ylims)

Small helper function then generates the xlims and ylims for the plot.

TaijaPlotting.get_names — Method

get_names(X)

Helper function to get variables names of X.

TaijaPlotting.plot_state — Method

plot_state(
    ce::CounterfactualExplanation,
    t::Int,
    final_state::Bool;
    kwargs...
)

Helper function that plots a single step of the counterfactual path.

TaijaPlotting.set_up_plots — Method

set_up_plots(
    ce::CounterfactualExplanation;
    alpha,
    plot_proba,
    kwargs...
)

A helper method that prepares data for plotting.